Refactoring: switch to java.time (#28)
* remove Date

part 1 of: delete Date and SimpleDateFormat

* remove SimpleDateFormat

and fix Tests

* refactor similar tests

* Update linter.yml

* Delete linter.yml

+++ src/test/java/net/aclrian/mpe/controller/TestFinishController.java
@@ -11,10 +11,8 @@ import net.aclrian.mpe.messdiener.*;
 import net.aclrian.mpe.messe.*;
 import net.aclrian.mpe.pfarrei.*;
 import net.aclrian.mpe.utils.*;
-import org.assertj.core.api.*;
 import org.junit.*;
 import org.mockito.*;
-import org.testfx.assertions.api.*;
 import org.testfx.assertions.api.Assertions;
 import org.testfx.framework.junit.*;
 import org.testfx.util.*;
@@ -22,6 +20,8 @@ import org.testfx.util.*;
 import java.io.*;
 import java.net.*;
 import java.nio.file.*;
+import java.time.*;
+import java.time.format.TextStyle;
 import java.util.*;
 
 public class TestFinishController extends ApplicationTest {
@@ -112,7 +112,7 @@ public class TestFinishController extends ApplicationTest {
         });
         WaitForAsyncUtils.waitForFxEvents();
 
-        Date date = TestFerienplanController.getToday();
+        LocalDate date = TestFerienplanController.getToday();
         m1.getMessdaten().einteilenVorzeitig(date, false);
 
 
@@ -169,152 +169,21 @@ public class TestFinishController extends ApplicationTest {
     }
 
     @Test
-    public void testEinteilung() {
-        Messdiener m1Freund = Mockito.mock(Messdiener.class);
-        Mockito.when(dv.getMessdiener()).thenReturn(Arrays.asList(m1, m2, m3, m1Freund));
-        Mockito.when(dv.getPfarrei()).thenReturn(pf);
-        StandartMesse standartMesse = new StandartMesse("So", 8, "00", "o1", 2, "t1");
-        Mockito.when(pf.getStandardMessen()).thenReturn(Collections.singletonList(standartMesse));
-        Mockito.when(pf.getSettings()).thenReturn(einst);
-        einst.editMaxDienen(false, 10);
-        einst.editMaxDienen(true, 10);
-        einst.editiereYear(0, 9);
-
-        Mockito.when(m1.toString()).thenReturn("m1");
-        Mockito.when(m2.toString()).thenReturn("m2");
-        Mockito.when(m3.toString()).thenReturn("m3");
-        final String m1FreundString = "dfc";
-        Mockito.when(m1Freund.toString()).thenReturn(m1FreundString);
-        Mockito.when(m1Freund.istLeiter()).thenReturn(false);
-        Mockito.when(m1.istLeiter()).thenReturn(false);
-        Mockito.when(m2.istLeiter()).thenReturn(true);
-        Mockito.when(m3.istLeiter()).thenReturn(false);
-
-        Mockito.when(m1.getEmail()).thenReturn("");
-        Mockito.when(m2.getEmail()).thenReturn("");
-        Mockito.when(m3.getEmail()).thenReturn("a@w.de");
-        Mockito.when(m1Freund.getEmail()).thenReturn("");
-
-        Mockito.when(m1.getGeschwister()).thenReturn(new String[0]);
-        Mockito.when(m2.getGeschwister()).thenReturn(new String[0]);
-        Mockito.when(m3.getGeschwister()).thenReturn(new String[0]);
-        Mockito.when(m1Freund.getGeschwister()).thenReturn(new String[0]);
-
-        Mockito.when(m1.getFreunde()).thenReturn(new String[]{m1FreundString});
-        Mockito.when(m2.getFreunde()).thenReturn(new String[0]);
-        Mockito.when(m3.getFreunde()).thenReturn(new String[0]);
-        Mockito.when(m1Freund.getFreunde()).thenReturn(new String[]{"m1"});
-
-        Mockito.when(m1.getEintritt()).thenReturn(Messdaten.getMaxYear());
-        Mockito.when(m2.getEintritt()).thenReturn(Messdaten.getMaxYear());
-        Mockito.when(m3.getEintritt()).thenReturn(Messdaten.getMaxYear());
-        Mockito.when(m1Freund.getEintritt()).thenReturn(Messdaten.getMaxYear());
-        Messdaten md1 = new Messdaten(m1);
-        Mockito.when(m1.getMessdaten()).thenReturn(md1);
-        Messdaten md2 = new Messdaten(m2);
-        Mockito.when(m2.getMessdaten()).thenReturn(md2);
-        Messdaten md3 = new Messdaten(m3);
-        Mockito.when(m3.getMessdaten()).thenReturn(md3);
-        Messdaten md1F = new Messdaten(m1Freund);
-        Mockito.when(m1Freund.getMessdaten()).thenReturn(md1F);
-        Messverhalten mv1 = new Messverhalten();
-        Messverhalten mv2 = new Messverhalten();
-        Messverhalten mv3 = new Messverhalten();
-        Messverhalten mv1F = new Messverhalten();
-
-        mv1.editiereBestimmteMesse(standartMesse, true);
-        md1.austeilen(TestFerienplanController.getYesterday2());
-        mv1F.editiereBestimmteMesse(standartMesse, true);
-        md1F.austeilen(TestFerienplanController.getYesterday2());
-
-        Mockito.when(m1.getDienverhalten()).thenReturn(mv1);
-        Mockito.when(m2.getDienverhalten()).thenReturn(mv2);
-        Mockito.when(m3.getDienverhalten()).thenReturn(mv3);
-        Mockito.when(m1Freund.getDienverhalten()).thenReturn(mv1F);
-
-        Messe me1 = new Messe(false, 1, TestFerienplanController.getYesterday2(), "o1", "t1");
-        Messe me2 = new Messe(TestFerienplanController.getToday(), standartMesse);
-        Messe me3 = new Messe(false, 1, TestFerienplanController.getTomorrow(), "o3", "t3");
-
-        Mockito.doCallRealMethod().when(dialogs).show(Mockito.anyList(), Mockito.eq(FinishController.NICHT_EINGETEILTE_MESSDIENER));
-
-        Assertions.assertThat(m1.getMessdaten().kann(me1.getDate(), false, false)).isFalse();
-        Assertions.assertThat(m1Freund.getMessdaten().kann(me1.getDate(), false, false)).isFalse();
-        Assertions.assertThat(m1.getMessdaten().kann(me2.getDate(), false, false)).isTrue();
-        Assertions.assertThat(m1Freund.getMessdaten().kann(me2.getDate(), false, false)).isTrue();
-        Platform.runLater(() -> {
-            URL u = getClass().getResource(MainController.EnumPane.PLAN.getLocation());
-            FXMLLoader fl = new FXMLLoader(u);
-            try {
-                List<Messe> messes = Arrays.asList(me1, me2, me3);
-                instance = new FinishController(null, messes);
-                fl.setController(instance);
-                pane.getChildren().add(fl.load());
-                instance.afterstartup(pane.getScene().getWindow(), mc);
-            } catch (IOException e) {
-                Log.getLogger().error(e.getMessage(), e);
-                Assertions.fail("Could not open " + MainController.EnumPane.FERIEN.getLocation() + e.getLocalizedMessage(), e);
-            }
-        });
-        WaitForAsyncUtils.waitForFxEvents();
-        Assertions.assertThat(me1.istFertig() && me2.istFertig() && me3.istFertig()).isTrue();
-        Assertions.assertThat(!me1.getEingeteilte().contains(m1) && !me1.getEingeteilte().contains(m1Freund)).isTrue();
-        Assertions.assertThat(!me3.getEingeteilte().contains(m1) && !me3.getEingeteilte().contains(m1Freund)).isTrue();
-        Assertions.assertThat(md1.kanndann(TestFerienplanController.getToday(), false)).isFalse();
-        Assertions.assertThat(md1F.kanndann(TestFerienplanController.getToday(), false)).isFalse();
-        Assertions.assertThat(me2.getEingeteilte().contains(m1) && me2.getEingeteilte().contains(m1)).isTrue();
-
-        File out = new File(Log.getWorkingDir().getAbsolutePath(), instance.getTitle() + ".pdf");
-        try {
-            if (out.exists()) {
-                Files.delete(out.toPath());
-            }
-            instance.toPDF(null);
-            Assertions.assertThat(out.exists()).isTrue();
-            Files.delete(out.toPath());
-        } catch (IOException e) {
-            Log.getLogger().error(e.getMessage(), e);
-            Assertions.fail(e.getMessage(), e);
-        }
-
-        Platform.runLater(() -> Mockito.when(dialogs.yesNoCancel(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString()))
-                .thenReturn(Dialogs.YesNoCancelEnum.YES));
-        WaitForAsyncUtils.waitForFxEvents();
-        Platform.runLater(() -> {
-            if (pane.lookup("#zurueck") instanceof Button zurueck) {
-                zurueck.fire();
-            } else {
-                Assertions.fail("Could not find zurueck");
-            }
-        });
-        WaitForAsyncUtils.waitForFxEvents();
-        for (Messdiener m : dv.getMessdiener()) {
-            if (m.getMessdaten().getInsgesamtEingeteilt() != 0 || m.getMessdaten().getAnzMessen() != 0) {
-                Assertions.fail("Not all Messdaten are zero");
-            }
-        }
-        Assertions.assertThat(me1.getEingeteilte().size()).isEqualTo(0);
-        Assertions.assertThat(me2.getEingeteilte().size()).isEqualTo(0);
-        Assertions.assertThat(me3.getEingeteilte().size()).isEqualTo(0);
-
-        Pair<List<Messdiener>, StringBuilder> pair = instance.getResourcesForEmail();
-        Assertions.assertThat(pair.getValue().toString()).matches("mailto:\\?bcc=a@w\\.de&subject=Messdienerplan%20vom%201[23].%20Juli%20bis%201[56]\\.%20Juli&body=%0D%0A");
-        try {
-            new URI(pair.getValue().toString());
-        } catch (URISyntaxException e) {
-            Assertions.fail(e.getMessage(), e);
-        }
-
-        Assertions.assertThat(pair.getKey()).hasSize(3).containsAll(Arrays.asList(m1, m2, m1Freund));
+    public void testEinteilungWithoutDOCX() {
+        testEinteilung(false);
     }
 
     @Ignore("Run Test with conversion to docx")
     @Test
-    public void testEinteilungWithConversionOfOutput() {
+    public void testEinteilungWithDOCX() {
+        testEinteilung(true);
+    }
+
+    private void testEinteilung(boolean skipDOCX) {
         Messdiener m1Freund = Mockito.mock(Messdiener.class);
         Mockito.when(dv.getMessdiener()).thenReturn(Arrays.asList(m1, m2, m3, m1Freund));
         Mockito.when(dv.getPfarrei()).thenReturn(pf);
-        StandartMesse standartMesse = new StandartMesse("Mo", 8, "00", "o1", 2, "t1");
+        StandartMesse standartMesse = new StandartMesse(DayOfWeek.MONDAY, 8, "00", "o1", 2, "t1");
         Mockito.when(pf.getStandardMessen()).thenReturn(Collections.singletonList(standartMesse));
         Mockito.when(pf.getSettings()).thenReturn(einst);
         einst.editMaxDienen(false, 10);
@@ -373,9 +242,9 @@ public class TestFinishController extends ApplicationTest {
         Mockito.when(m3.getDienverhalten()).thenReturn(mv3);
         Mockito.when(m1Freund.getDienverhalten()).thenReturn(mv1F);
 
-        Messe me1 = new Messe(false, 1, TestFerienplanController.getYesterday2(), "o1", "t1");
-        Messe me2 = new Messe(TestFerienplanController.getToday(), standartMesse);
-        Messe me3 = new Messe(false, 1, TestFerienplanController.getTomorrow(), "o3", "t3");
+        Messe me1 = new Messe(false, 1, TestFerienplanController.getYesterday2().atTime(0, 0), "o1", "t1");
+        Messe me2 = new Messe(TestFerienplanController.getToday().atTime(0, 0), standartMesse);
+        Messe me3 = new Messe(false, 1, TestFerienplanController.getTomorrow().atTime(0, 0), "o3", "t3");
 
         Mockito.doCallRealMethod().when(dialogs).show(Mockito.anyList(), Mockito.eq(FinishController.NICHT_EINGETEILTE_MESSDIENER));
 
@@ -417,17 +286,19 @@ public class TestFinishController extends ApplicationTest {
             Log.getLogger().error(e.getMessage(), e);
             Assertions.fail(e.getMessage(), e);
         }
-        out = new File(Log.getWorkingDir(), instance.getTitle() + ".docx");
-        try {
-            if (out.exists()) {
+        if(skipDOCX) {
+            out = new File(Log.getWorkingDir(), instance.getTitle() + ".docx");
+            try {
+                if (out.exists()) {
+                    Files.delete(out.toPath());
+                }
+                instance.toWORD(null);
+                Assertions.assertThat(out.exists()).isTrue();
                 Files.delete(out.toPath());
+            } catch (IOException e) {
+                Log.getLogger().error(e.getMessage(), e);
+                Assertions.fail(e.getMessage(), e);
             }
-            instance.toWORD(null);
-            Assertions.assertThat(out.exists()).isTrue();
-            Files.delete(out.toPath());
-        } catch (IOException e) {
-            Log.getLogger().error(e.getMessage(), e);
-            Assertions.fail(e.getMessage(), e);
         }
 
         Platform.runLater(() -> Mockito.when(dialogs.yesNoCancel(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString()))
@@ -450,8 +321,12 @@ public class TestFinishController extends ApplicationTest {
         Assertions.assertThat(me2.getEingeteilte().size()).isEqualTo(0);
         Assertions.assertThat(me3.getEingeteilte().size()).isEqualTo(0);
 
+        int from = TestFerienplanController.getYesterday2().getDayOfMonth();
+        int to = TestFerienplanController.getTomorrow().getDayOfMonth();
+        String fromMonth = TestFerienplanController.getYesterday2().getMonth().getDisplayName(TextStyle.FULL, Locale.getDefault());
+        String toMonth = TestFerienplanController.getTomorrow().getMonth().getDisplayName(TextStyle.FULL, Locale.getDefault());
         Pair<List<Messdiener>, StringBuilder> pair = instance.getResourcesForEmail();
-        Assertions.assertThat(pair.getValue().toString()).isEqualTo("mailto:?bcc=a@w.de&subject=Messdienerplan%20vom%2012.%20Juli%20bis%2015.%20Juli&body=%0D%0A");
+        Assertions.assertThat(pair.getValue().toString()).isEqualTo("mailto:?bcc=a@w.de&subject=Messdienerplan%20vom%20"+from+".%20"+fromMonth+"%20bis%20"+to+".%20"+toMonth+"&body=%0D%0A");
         try {
             new URI(pair.getValue().toString());
         } catch (URISyntaxException e) {
